Moose: Make it build on macOS 10.14 and Xcode 10. Note that currently shortcut input and launch at login are broken. Many deprecation warnings, too. But at least it builds and runs again.

+++ TalkingMoose/UKMoosePhraseologyDelegate.m
@@ -1,9 +1,10 @@
 #import "UKMoosePhraseologyDelegate.h"
 #import "UKGroupFile.h"
+#import "UKHelperMacros.h"
 
 @implementation UKMoosePhraseologyDelegate
 
-int		UKPhraseFileSortFunction( id objA, id objB, void* context )
+NSInteger	UKPhraseFileSortFunction( id objA, id objB, void* context )
 {
 	return [[objA objectForKey: @"name"] caseInsensitiveCompare: [objB objectForKey: @"name"]];
 }
